**Major Upgrade and Enhancements**

This relese includes a major upgrade with numerous bug fixes and new features, significantly improving the overall game engine implementation, which now sits at 87%. completed.

**Key Highlights:**

- **Bug Fixes:**
  - Resolved multiple issues, most notably in the sound system when restoring from savegame.
  - Fixed GDI dialogs on Windows 11.
  - Corrected drawing of overlay maps.

- **New Features:**
  - Introduced developer console commands.
  - Implemented parsing and compiling of COG scripts.
  - Added developer COG host functions.
  - 
- **Enhancements:**
  - Added support for QWXGA+ resolutions (QHD, 4K, etc.).
  - Improved AI movement at cliff edges.
  - Enhanced overall stability and performance.
  - Improved user experience with various quality-of-life updates.

+++ Jones3D/Main/JonesMain.h
@@ -50,7 +50,7 @@ void JonesMain_RefreshDisplayDevice(void);
 
 int JonesMain_PlayIntroMovie(void);
 
-void J3DAPI JonesMain_BindToggleMenuControlKeys(const int* paKeyIds, int numKeys);
+void J3DAPI JonesMain_BindToggleMenuControlKeys(const size_t* paKeyIds, size_t numKeys);
 
 void J3DAPI JonesMain_LogErrorToFile(const char* pErrorText);
 J3DNORETURN void JonesMain_FatalErrorUnknown(void); // Added
